Add support for dots and ticks in lines.
Internally, this works by creating a new point type, but treating the dot
or tick like an edge. When the line is cleared, the tick is replaced as a
corner as to not ruin the display of other lines that may intersect with
it. Additionally, we only ever need a single tick object to display at
this point.