Fix issues around ticks / dots without intersecting lines.
If no other lines touch a tick or dot, when we change the tick to
a point, it will not be recognized as such. In this case, we will
leave the resulting '+' as text behind the tick or dot, which looks
odd. Since this essentially becomes a point, add it so that the
'+' is cleared before we do text parsing.