Fix off-by-one.
This off-by-one causes lines to start or end outside of their boundaries.
This happens because our last call to the function is necessarily outside
of the bounds of the line. When we find that we have completed the line,
we should add the *last valid* point, not the point upon which we
terminate.

This greatly improves many of the 0MQ book diagrams.